refactor: ♻️ encapsulate watch button
